requirements.txt
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
schedule==1.2.1
